refactor(mapper): move map logic to separate map component
